Compliant with tests directory as previous created
